refactor: correct route
