+++ src/App.js
@@ -3,7 +3,7 @@ import Header from './components/Header/Header';
 import Footer from './components/Footer/Footer';
 import {BrowserRouter} from "react-router-dom";
 import React, {useState} from "react";
-import AppRouter from "./components/AppRouter";
+import AppRouter from "./components/RouterComp/AppRouter";
 import {AuthContext} from "./context";
 
 function App() {
