+++ src/components/pages/Authorization/Authorization.jsx
@@ -3,11 +3,13 @@ import cl from './Authorization.module.css'
 import Input from "../../UI/Input/Input";
 import Button from "../../UI/Button/Button";
 import {AuthContext} from "../../../context";
-import {useNavigate} from "react-router-dom";
+import {useLocation, useNavigate} from "react-router-dom";
 
 const Authorization = () => {
     const {setUserId} = useContext(AuthContext);
     const navigate = useNavigate();
+    const location = useLocation();
+    const from = location.state?.from?.pathname || '/';
 
     const login = (event) => {
         event.preventDefault();
@@ -15,7 +17,7 @@ const Authorization = () => {
         const userId = '0ff617b5-3691-4e20-b58b-7e91740aed06';
         setUserId(userId);
         localStorage.setItem('auth', userId);
-        navigate('/home');
+        navigate(from);
     }
 
     return (
