+++ src/router/index.js
@@ -1,7 +1,6 @@
 import Home from "../components/pages/Home/Home";
 import PersonalAccount from "../components/pages/PersonalAccount/PersonalAccount";
 import Error from "../components/pages/Error/Error";
-import Authorization from "../components/pages/Authorization/Authorization";
 import {Navigate} from "react-router-dom";
 
 export const privateRoutes = [
@@ -12,7 +11,10 @@ export const privateRoutes = [
     {path: '*', element: <Navigate to="/home" replace />},
 ];
 
-export const publicRoutes = [
-    {path: '/authorization', element: <Authorization/>},
-    {path: '*', element: <Navigate to="/authorization" replace />},
+export const adminRoutes = [
+    {path: '/', element: <Navigate to="/home" replace />},
+    {path: '/home', element: <Home/>},
+    {path: '/personal-account', element: <PersonalAccount/>},
+    {path: 'error', element: <Error/>},
+    {path: '*', element: <Navigate to="/home" replace />},
 ];
